+++ Laba_1/2.cpp
@@ -4,7 +4,7 @@ int main() {
     int a = 3;
     int b = 4;
 
-    double c = sqrt(a*a + b*b);
+    float c = sqrt(a*a + b*b);
 
     std::cout << "c = " << c << std::endl;
