+++ Laba_1/8.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 
 int main() {
-    double x1, y1, x2, y2;
+    float x1, y1, x2, y2;
     std::cout << "Введите x1: ";
     std::cin >> x1;
     std::cout << "Введите y1: ";
@@ -12,7 +12,7 @@ int main() {
     std::cout << "Введите y2: ";
     std::cin >> y2;
 
-    double distance = std::sqrt(std::pow(x2 - x1, 2) + std::pow(y2 - y1, 2));
+    float distance = std::sqrt(std::pow(x2 - x1, 2) + std::pow(y2 - y1, 2));
 
     std::cout << "Расстояние между точками: " << distance << std::endl;
